# contracts/iot-ingestion.openapi.yaml
openapi: 3.1.0

info:
  title: IoT Ingestion Service API
  version: 1.0.0
  description: |
    API contract cho IoT Ingestion Service trong hệ thống Smart Campus.

    Service này nhận telemetry readings từ thiết bị IoT và cung cấp endpoint
    để lấy đọc số mới nhất.

  contact:
    name: Trần Công Quân
    email: congquan@example.com

servers:
  - url: http://localhost:4010
    description: Prism mock server

  - url: http://localhost:8000
    description: Local service endpoint

tags:
  - name: health
    description: Kiểm tra trạng thái service

  - name: readings
    description: API đọc dữ liệu telemetry

security:
  - bearerAuth: []

paths:

  /health:
    get:
      operationId: getHealth
      summary: Kiểm tra trạng thái service
      description: Endpoint kiểm tra IoT Ingestion Service hoạt động.
      tags:
        - health
      security: []
      responses:
        '200':
          description: Service đang hoạt động
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'

  /readings:
    post:
      operationId: createReading
      summary: Gửi một reading IoT mới
      description: Nhận reading telemetry từ thiết bị IoT.
      tags:
        - readings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReadingInput'
      responses:
        '201':
          description: Reading được chấp nhận
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadingAccepted'
              examples:
                success:
                  value:
                    reading_id: rdg-1234
                    accepted: true
                    received_at: '2026-05-13T08:30:00+07:00'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /readings/latest:
    get:
      operationId: getLatestReadings
      summary: Lấy latest readings
      description: Trả về danh sách latest readings theo thiết bị.
      tags:
        - readings
      parameters:
        - name: device_id
          in: query
          required: false
          description: Mã thiết bị để lọc readings
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: Số lượng reading cần lấy
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 5
      responses:
        '200':
          description: Danh sách latest readings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadingList'
              examples:
                success:
                  value:
                    items:
                      - reading_id: rdg-1234
                        device_id: ESP32-LAB-A01
                        metric: temperature
                        value: 31.5
                        unit: celsius
                        timestamp: '2026-05-13T08:30:00+07:00'
                      - reading_id: rdg-1235
                        device_id: ESP32-LAB-A01
                        metric: humidity
                        value: 55.2
                        unit: percent
                        timestamp: '2026-05-13T08:31:00+07:00'
                    total: 2
                    limit: 2
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      additionalProperties: false
      required:
        - status
        - service
        - time
      properties:
        status:
          type: string
          example: ok
        service:
          type: string
          example: iot-ingestion-service
        time:
          type: string
          format: date-time

    ReadingInput:
      type: object
      additionalProperties: false
      required:
        - device_id
        - metric
        - value
        - unit
        - timestamp
      properties:
        device_id:
          type: string
          example: ESP32-LAB-A01
        metric:
          type: string
          enum: [temperature, humidity, motion, smoke]
          example: temperature
        value:
          type: number
          example: 31.5
        unit:
          type: string
          enum: [celsius, percent, boolean, ppm]
          example: celsius
        timestamp:
          type: string
          format: date-time
          example: '2026-05-13T08:30:00+07:00'

    ReadingAccepted:
      type: object
      additionalProperties: false
      required:
        - reading_id
        - accepted
      properties:
        reading_id:
          type: string
          example: rdg-1234
        accepted:
          type: boolean
          example: true
        received_at:
          type: string
          format: date-time
          example: '2026-05-13T08:30:00+07:00'

    ReadingSummary:
      type: object
      additionalProperties: false
      required:
        - reading_id
        - device_id
        - metric
        - value
        - unit
        - timestamp
      properties:
        reading_id:
          type: string
          example: rdg-1234
        device_id:
          type: string
          example: ESP32-LAB-A01
        metric:
          type: string
          example: temperature
        value:
          type: number
          example: 31.5
        unit:
          type: string
          example: celsius
        timestamp:
          type: string
          format: date-time
          example: '2026-05-13T08:30:00+07:00'

    ReadingList:
      type: object
      additionalProperties: false
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReadingSummary'
        total:
          type: integer
          example: 2
        limit:
          type: integer
          example: 2

    ProblemDetails:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri

  responses:
    BadRequest:
      description: Payload sai định dạng
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            invalidPayload:
              value:
                type: 'https://example.com/probs/invalid-payload'
                title: 'Invalid payload'
                status: 400
                detail: 'Thiếu trường bắt buộc hoặc định dạng sai'
    Unauthorized:
      description: Thiếu Bearer token
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            missingAuth:
              value:
                type: 'https://example.com/probs/unauthorized'
                title: 'Unauthorized'
                status: 401
                detail: 'Thiếu token hoặc token không hợp lệ'
    NotFound:
      description: Resource không tồn tại
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Conflict:
      description: Xung đột dữ liệu
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    UnprocessableEntity:
      description: Payload đúng JSON nhưng sai nghiệp vụ
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    TooManyRequests:
      description: Quá nhiều request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    InternalServerError:
      description: Lỗi hệ thống phía IoT Ingestion
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'